move Track control components into directory
