+++ src/util/device-id-from-stream.ts
