+++ src/util/get-latency-samples.ts
@@ -1,4 +1,4 @@
-import { logger } from '../util/logger'
+import { logger } from './logger'
 
 /**
  *
